rename: move home frame to core from profile
