+++ src/features/auth/frames/register-frame.tsx
@@ -1,7 +1,8 @@
-import { postUser } from '@/data/post-user';
-import { Pretty, RequireKeys, useFunnel } from '@/features/core';
 import { useNavigate } from '@tanstack/react-router';
 import { overlay } from 'overlay-kit';
+
+import { postUser } from '@/data/post-user';
+import { Pretty, RequireKeys, useFunnel } from '@/features/core';
 
 import { RegisterStepUndoWarningOverlay } from '../components/register-step-undo-warning-overlay';
 import { CodeStep } from './register-steps/code-step';
