+++ src/features/core/frames/change-password-frame.tsx
@@ -1,4 +1,5 @@
 import { useEffect } from 'react';
+
 import toast from 'react-hot-toast';
 import { useTranslation } from 'react-i18next';
