+++ src/features/profile/components/profile-edit-overlay.tsx
@@ -15,7 +15,7 @@ import {
 } from '@/features/core';
 import { useLoading } from '@/features/core';
 
-import { useProfileEditForm } from '../hooks/use-profile-edit-form';
+import { useProfileEditForm } from '../../core/hooks/use-profile-edit-form';
 
 export function ProfileEditOverlay({
   isOpen,
