+++ src/features/profile/index.ts
@@ -1,4 +1,1 @@
-export * from './frames/home-frame';
 export * from './frames/profile-frame';
-export * from './frames/change-password-frame';
-export * from './frames/issue-password-frame';
